move NestedWizard logic to Wizzard class
