+++ packages/core/src/shared/wizards/stateController.ts
@@ -16,7 +16,7 @@ export enum ControlSignal {
  * Value for indicating current direction of the wizard
  * Created mainly to support skipping prompters
  */
-export enum DIRECTION {
+export enum Direction {
     Forward,
     Backward,
 }
@@ -48,11 +48,11 @@ export class StateMachineController<TState> {
     private extraSteps = new Map<number, Branch<TState>>()
     private steps: Branch<TState> = []
     private internalStep: number = 0
-    private direction: DIRECTION
+    private direction: Direction
 
     public constructor(private state: TState = {} as TState) {
         this.previousStates = [_.cloneDeep(state)]
-        this.direction = DIRECTION.Forward
+        this.direction = Direction.Forward
     }
 
     public addStep(step: StepFunction<TState>): void {
@@ -82,14 +82,14 @@ export class StateMachineController<TState> {
 
         this.state = this.previousStates.pop()!
         this.internalStep -= 1
-        this.direction = DIRECTION.Backward
+        this.direction = Direction.Backward
     }
 
     protected advanceState(nextState: TState): void {
         this.previousStates.push(this.state)
         this.state = nextState
         this.internalStep += 1
-        this.direction = DIRECTION.Forward
+        this.direction = Direction.Forward
     }
 
     protected detectCycle(step: StepFunction<TState>): TState | undefined {
@@ -122,7 +122,7 @@ export class StateMachineController<TState> {
                 /**
                  * Depending on current wizard direction, skip signal get converted to forward or backward control signal
                  */
-                result.controlSignal = this.direction === DIRECTION.Forward ? undefined : ControlSignal.Back
+                result.controlSignal = this.direction === Direction.Forward ? undefined : ControlSignal.Back
             }
             return result
         } else {
